--- docs/pages/reference/account-kit/core/classes/ClientOnlyPropertyError/constructor.mdx
@@ -1,26 +1,0 @@
----
-# This file is autogenerated
-title: ClientOnlyPropertyError
-description: Overview of the ClientOnlyPropertyError method
-slug: wallets/reference/account-kit/core/classes/ClientOnlyPropertyError/constructor
----
-
-Creates a new ClientOnlyPropertyError
-
-<Note>
-  `ClientOnlyPropertyError` extends `BaseError`, see the docs for BaseError for
-  all supported methods.
-</Note>
-
-## Import
-
-```ts
-import { ClientOnlyPropertyError } from "@account-kit/core";
-```
-
-## Parameters
-
-### property
-
-`string`
-the name of the property that is only available on the client
