--- docs/pages/reference/account-kit/core/functions/convertSignerStatusToState.mdx
@@ -1,31 +1,0 @@
----
-# This file is autogenerated
-title: convertSignerStatusToState
-description: Overview of the convertSignerStatusToState method
-slug: wallets/reference/account-kit/core/functions/convertSignerStatusToState
----
-
-Converts the AlchemySigner's status to a more readable object
-
-## Import
-
-```ts
-import { convertSignerStatusToState } from "@account-kit/core";
-```
-
-## Parameters
-
-### alchemySignerStatus
-
-`AlchemySignerStatus`
-Enum value of the AlchemySigner's status to convert
-
-### error
-
-`ErrorInfo | undefined`
-the current signer error, if present
-
-## Returns
-
-`SignerStatus`
-an object containing the original status as well as booleans to check the current state
