--- docs/pages/reference/account-kit/core/functions/cookieToInitialState.mdx
@@ -1,31 +1,0 @@
----
-# This file is autogenerated
-title: cookieToInitialState
-description: Overview of the cookieToInitialState method
-slug: wallets/reference/account-kit/core/functions/cookieToInitialState
----
-
-Converts a cookie into an initial state object
-
-## Import
-
-```ts
-import { cookieToInitialState } from "@account-kit/core";
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config containing the client store
-
-### cookie
-
-`string | undefined`
-optional cookie string
-
-## Returns
-
-`StoredState | undefined`
-the deserialized AlchemyClientState if the cookie exists, otherwise undefined
