--- docs/pages/reference/account-kit/core/functions/createAccount.mdx
@@ -1,46 +1,0 @@
----
-# This file is autogenerated
-title: createAccount
-description: Overview of the createAccount method
-slug: wallets/reference/account-kit/core/functions/createAccount
----
-
-Creates an account of a specified type using the provided parameters and configuration. Supports creating LightAccount and MultiOwnerModularAccount types.
-
-## Import
-
-```ts
-import { createAccount } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { createAccount } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-const account = createAccount(
-  {
-    type: "LightAccount",
-  },
-  config,
-);
-```
-
-## Parameters
-
-### params
-
-`CreateAccountParams<TAccount>`
-The parameters required to create the account, including the type and account parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration object for Alchemy accounts
-
-## Returns
-
-`Promise<SupportedAccounts>`
-A promise that resolves to the created account object
