--- docs/pages/reference/account-kit/core/functions/createConfig.mdx
@@ -1,42 +1,0 @@
----
-# This file is autogenerated
-title: createConfig
-description: Overview of the createConfig method
-slug: wallets/reference/account-kit/core/functions/createConfig
----
-
-Creates an AlchemyAccountsConfig object that can be used in conjunction with
-the actions exported from `@account-kit/core`.
-
-The config contains core and client stores that can be used to manage account state
-in your application.
-
-## Import
-
-```ts
-import { createConfig } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { createConfig } from "@account-kit/core";
-import { sepolia } from "@account-kit/infra";
-
-const config = createConfig({
-  chain: sepolia,
-  transport: alchemy({ apiKey: "your-api-key" }),
-});
-```
-
-## Parameters
-
-### params
-
-`CreateConfigProps`
-The parameters to create the config with
-
-## Returns
-
-`AlchemyAccountsConfig`
-An alchemy account config object containing the core and client store
