--- docs/pages/reference/account-kit/core/functions/createDefaultAccountState.mdx
@@ -1,26 +1,0 @@
----
-# This file is autogenerated
-title: createDefaultAccountState
-description: Overview of the createDefaultAccountState method
-slug: wallets/reference/account-kit/core/functions/createDefaultAccountState
----
-
-Creates the default account state for the given chains.
-
-## Import
-
-```ts
-import { createDefaultAccountState } from "@account-kit/core";
-```
-
-## Parameters
-
-### chains
-
-`Chain[]`
-The chains to create the account state for
-
-## Returns
-
-`NoUndefined<StoreState["accounts"]>`
-The default account state for the given chains
