--- docs/pages/reference/account-kit/core/functions/defaultAccountState.mdx
@@ -1,27 +1,0 @@
----
-# This file is autogenerated
-title: defaultAccountState
-description: Overview of the defaultAccountState method
-slug: wallets/reference/account-kit/core/functions/defaultAccountState
----
-
-Returns the default state for an account of a supported type.
-
-## Import
-
-```ts
-import { defaultAccountState } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { defaultAccountState } from "@account-kit/core";
-
-const defaultLightAccountState = defaultAccountState<"LightAccount">();
-```
-
-## Returns
-
-`AccountState<T>`
-The default state for the specified account type
