--- docs/pages/reference/account-kit/core/functions/disconnect.mdx
@@ -1,35 +1,0 @@
----
-# This file is autogenerated
-title: disconnect
-description: Overview of the disconnect method
-slug: wallets/reference/account-kit/core/functions/disconnect
----
-
-Disconnects the current signer, accounts, and clears the store.
-
-## Import
-
-```ts
-import { disconnect } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { disconnect, createConfig } from "@account-kit/core";
-import { sepolia } from "@account-kit/infra";
-
-const config = createConfig({
-  chain: sepolia,
-  apiKey: "your-api-key",
-});
-
-await disconnect(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration containing the store to be cleared
