--- docs/pages/reference/account-kit/core/functions/getBundlerClient.mdx
@@ -1,35 +1,0 @@
----
-# This file is autogenerated
-title: getBundlerClient
-description: Overview of the getBundlerClient method
-slug: wallets/reference/account-kit/core/functions/getBundlerClient
----
-
-Retrieves the BundlerClient from the core store of the given AlchemyAccountsConfig.
-
-## Import
-
-```ts
-import { getBundlerClient } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-// see `createConfig` for more information on how to create a config
-import { config } from "./config";
-
-const bundlerClient = getBundlerClient(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration object containing the core store.
-
-## Returns
-
-`ClientWithAlchemyMethods`
-The BundlerClient from the core store.
