--- docs/pages/reference/account-kit/core/functions/getChain.mdx
@@ -1,26 +1,0 @@
----
-# This file is autogenerated
-title: getChain
-description: Overview of the getChain method
-slug: wallets/reference/account-kit/core/functions/getChain
----
-
-Gets the currently active chain
-
-## Import
-
-```ts
-import { getChain } from "@account-kit/core";
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config object
-
-## Returns
-
-`Chain`
-the currently active chain
