--- docs/pages/reference/account-kit/core/functions/getConnection.mdx
@@ -1,35 +1,0 @@
----
-# This file is autogenerated
-title: getConnection
-description: Overview of the getConnection method
-slug: wallets/reference/account-kit/core/functions/getConnection
----
-
-Used to get the connection for the currently active chain
-
-## Import
-
-```ts
-import { getConnection } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { getConnection } from "@account-kit/core";
-import { config } from "./config";
-
-const connection = getConnection(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config
-
-## Returns
-
-`Connection`
-a connection object for the current active chain
