--- docs/pages/reference/account-kit/core/functions/getSigner.mdx
@@ -1,37 +1,0 @@
----
-# This file is autogenerated
-title: getSigner
-description: Overview of the getSigner method
-slug: wallets/reference/account-kit/core/functions/getSigner
----
-
-If there is a signer attached to the client state, it will return it.
-The signer should always be null on the server, and will be set on the client
-if the store was properly hydrated.
-
-## Import
-
-```ts
-import { getSigner } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { getSigner } from "@account-kit/core";
-import { config } from "./config";
-
-const signer = getSigner(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The account config which contains the client store
-
-## Returns
-
-`AlchemySigner | null`
-the instance of the signer present in the store if it exists, otherwise null
