--- docs/pages/reference/account-kit/core/functions/getSignerStatus.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: getSignerStatus
-description: Overview of the getSignerStatus method
-slug: wallets/reference/account-kit/core/functions/getSignerStatus
----
-
-Retrieves the signer status from the client's store in the provided configuration.
-
-## Import
-
-```ts
-import { getSignerStatus } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { getSignerStatus } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-const signerStatus = getSignerStatus(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration object containing the client store
-
-## Returns
-
-`SignerStatus`
-The current signer status from the client store
