--- docs/pages/reference/account-kit/core/functions/getSolanaConnection.mdx
@@ -1,35 +1,0 @@
----
-# This file is autogenerated
-title: getSolanaConnection
-description: Overview of the getSolanaConnection method
-slug: wallets/reference/account-kit/core/functions/getSolanaConnection
----
-
-Used to get the connection for the id
-
-## Import
-
-```ts
-import { getSolanaConnection } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { getSolanaConnection } from "@account-kit/core";
-import { config } from "./config";
-
-const connection = getSolanaConnection(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config
-
-## Returns
-
-`SolanaConnection | void`
-a connection object for the current active chain
