--- docs/pages/reference/account-kit/core/functions/getUser.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: getUser
-description: Overview of the getUser method
-slug: wallets/reference/account-kit/core/functions/getUser
----
-
-Returns the currently logged in user if using an SCA with the AlchemySigner
-or the connected EOA details.
-
-## Import
-
-```ts
-import { getUser } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { getUser } from "@account-kit/core";
-import { config } from "./config";
-
-const user = getUser(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config containing app state
-
-## Returns
-
-`GetUserResult`
-the user if the signer or an EOA are connected
