--- docs/pages/reference/account-kit/core/functions/parseCookie.mdx
@@ -1,31 +1,0 @@
----
-# This file is autogenerated
-title: parseCookie
-description: Overview of the parseCookie method
-slug: wallets/reference/account-kit/core/functions/parseCookie
----
-
-Helper function that can be used to parse a cookie string on the server or client
-
-## Import
-
-```ts
-import { parseCookie } from "@account-kit/core";
-```
-
-## Parameters
-
-### cookie
-
-`string`
-the cookie string to parse
-
-### key
-
-`string`
-the key of the cookie to parse
-
-## Returns
-
-`string`
-the value of the cookie given a key if it exists, otherwise undefined
