--- docs/pages/reference/account-kit/core/functions/reconnect.mdx
@@ -1,31 +1,0 @@
----
-# This file is autogenerated
-title: reconnect
-description: Overview of the reconnect method
-slug: wallets/reference/account-kit/core/functions/reconnect
----
-
-This method will use the current state in the client store and attempt to restore
-connected instances of previously used accounts and the signer.
-
-## Import
-
-```ts
-import { reconnect } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { reconnect } from "@account-kit/core";
-import { config } from "./config";
-
-await reconnect(config);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config which contains the client store
