--- docs/pages/reference/account-kit/core/functions/watchAccount.mdx
@@ -1,41 +1,0 @@
----
-# This file is autogenerated
-title: watchAccount
-description: Overview of the watchAccount method
-slug: wallets/reference/account-kit/core/functions/watchAccount
----
-
-Watches for changes to a specific type of account and triggers the provided callback function when changes occur.
-
-## Import
-
-```ts
-import { watchAccount } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchAccount } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchAccount("LightAccount", config)(console.log);
-```
-
-## Parameters
-
-### type
-
-`TAccount`
-The type of account to watch
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration containing client store settings
-
-## Returns
-
-`(onChange: (account: GetAccountResult<TAccount>) => void) => (() => void)`
-A function that accepts a callback to be called when the account changes and returns a function to unsubscribe from the store
