--- docs/pages/reference/account-kit/core/functions/watchBundlerClient.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchBundlerClient
-description: Overview of the watchBundlerClient method
-slug: wallets/reference/account-kit/core/functions/watchBundlerClient
----
-
-Watches for changes to the bundler client within the given configuration and triggers a callback when changes occur.
-
-## Import
-
-```ts
-import { watchBundlerClient } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchBundlerClient } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchBundlerClient(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-The configuration object containing the core store
-
-## Returns
-
-`(onChange: (bundlerClient: ClientWithAlchemyMethods) => void) => (() => void)`
-A function accepting a callback function to invoke when the bundler client changes and returns a function to unsubscribe from the store
