--- docs/pages/reference/account-kit/core/functions/watchChain.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchChain
-description: Overview of the watchChain method
-slug: wallets/reference/account-kit/core/functions/watchChain
----
-
-Allows you to subscribe to changes of the chain in the client store.
-
-## Import
-
-```ts
-import { watchChain } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchChain } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchChain(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config object
-
-## Returns
-
-`(onChange: (chain: Chain) => void) => (() => void)`
-a function which accepts an onChange callback that will be fired when the chain changes and returns a function to unsubscribe from the store
