--- docs/pages/reference/account-kit/core/functions/watchConnection.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchConnection
-description: Overview of the watchConnection method
-slug: wallets/reference/account-kit/core/functions/watchConnection
----
-
-Subscribe to changes to the active connection
-
-## Import
-
-```ts
-import { watchConnection } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchConnection } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchConnection(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config
-
-## Returns
-
-`(onChange: (connection: Connection) => void) => (() => void)`
-a function which accepts an onChange callback that will be fired when the connection changes and returns a function to unsubscribe from the store
