--- docs/pages/reference/account-kit/core/functions/watchSigner.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchSigner
-description: Overview of the watchSigner method
-slug: wallets/reference/account-kit/core/functions/watchSigner
----
-
-Subscribe to changes of the signer instance on the client store.
-
-## Import
-
-```ts
-import { watchSigner } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchSigner } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchSigner(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config containing the client store
-
-## Returns
-
-`(onChange: (chain: AlchemySigner) => void) => (() => void)`
-a function which accepts an onChange callback that will be fired when the signer changes and returns a function to unsubscribe from the store
