--- docs/pages/reference/account-kit/core/functions/watchSolanaConnection.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchSolanaConnection
-description: Overview of the watchSolanaConnection method
-slug: wallets/reference/account-kit/core/functions/watchSolanaConnection
----
-
-Subscribe to changes to the solana connection for the id
-
-## Import
-
-```ts
-import { watchSolanaConnection } from "@account-kit/core";
-```
-
-## Usage
-
-```ts twoslash
-import { watchSolanaConnection } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-const config = {} as any;
-
-watchSolanaConnection(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the account config of the connection
-
-## Returns
-
-`(onChange: (connection: Connection) => void) => (() => void)`
-a function which accepts an onChange callback that will be fired when the connection changes and returns a function to unsubscribe from the store
