--- docs/pages/reference/account-kit/core/functions/watchUser.mdx
@@ -1,36 +1,0 @@
----
-# This file is autogenerated
-title: watchUser
-description: Overview of the watchUser method
-slug: wallets/reference/account-kit/core/functions/watchUser
----
-
-Watches for changes to the user in the client store and triggers the provided callback when a change is detected.
-
-## Import
-
-```ts
-import { watchUser } from "@account-kit/core";
-```
-
-## Usage
-
-```ts
-import { watchUser } from "@account-kit/core";
-// see createConfig for more information on how to create a config
-import { config } from "./config";
-
-watchUser(config)(console.log);
-```
-
-## Parameters
-
-### config
-
-`AlchemyAccountsConfig`
-the configuration containing the client store
-
-## Returns
-
-`(onChange: (user: User) => void) => (() => void)`
-a function which accepts a callback that fires when the user changes and returns a function to unsubscribe from the user updates
