+++ docs/pages/reference/account-kit/infra/functions/alchemyActions.mdx
@@ -1,0 +1,36 @@
+---
+# This file is autogenerated
+title: alchemyActions
+description: Overview of the alchemyActions method
+slug: wallets/reference/account-kit/infra/functions/alchemyActions
+---
+
+Provides a set of actions for interacting with the Alchemy Smart Account client, including the ability to simulate user operations.
+
+## Import
+
+```ts
+import { alchemyActions } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { alchemyActions } from "@account-kit/infra";
+import { createPublicClient } from "viem";
+
+const client = createPublicClient(...);
+const clientWithAlchemyActions = client.extend(alchemyActions);
+```
+
+## Parameters
+
+### client\_
+
+`Client<TTransport, TChain, TAccount>`
+The client instance used to perform actions
+
+## Returns
+
+`AlchemySmartAccountClientActions<TAccount, TContext>`
+An object containing Alchemy Smart Account client actions
