+++ docs/pages/reference/account-kit/infra/functions/alchemyEnhancedApiActions.mdx
@@ -1,0 +1,38 @@
+---
+# This file is autogenerated
+title: alchemyEnhancedApiActions
+description: Overview of the alchemyEnhancedApiActions method
+slug: wallets/reference/account-kit/infra/functions/alchemyEnhancedApiActions
+---
+
+Given an instance of the Alchemy SDK, returns a smart account client decorator which contains actions for interacting Alchemy's enhanced APIs.
+
+## Import
+
+```ts
+import { alchemyEnhancedApiActions } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { Alchemy } from "alchemy-sdk";
+import { alchemyEnhancedApiActions } from "@account-kit/infra";
+import { alchemySCAClient } from "./client";
+
+const alchemy = new Alchemy(...);
+const enhancedApiDecorator = alchemyEnhancedApiActions(alchemy);
+const withEnhancedApis = alchemySCAClient.extend(enhancedApiDecorator);
+```
+
+## Parameters
+
+### alchemy
+
+`Alchemy`
+The Alchemy instance containing the SDK client
+
+## Returns
+
+`(client: AlchemySmartAccountClient) => AlchemyEnhancedApis`
+A client decorator for Alchemy Smart Account clients that adds the enhanced API methods
