+++ docs/pages/reference/account-kit/infra/functions/alchemyFeeEstimator.mdx
@@ -1,0 +1,44 @@
+---
+# This file is autogenerated
+title: alchemyFeeEstimator
+description: Overview of the alchemyFeeEstimator method
+slug: wallets/reference/account-kit/infra/functions/alchemyFeeEstimator
+---
+
+Function that estimates the transaction fees using Alchemy methods for a given client.
+It fetches the latest block and estimates the max priority fee per gas, applying any overrides or fee options provided.
+
+## Import
+
+```ts
+import { alchemyFeeEstimator } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { alchemyFeeEstimator, alchemy } from "@account-kit/infra";
+import { createSmartAccountClient } from "@aa-sdk/core";
+
+const alchemyTransport = alchemy({
+  chain: sepolia,
+  apiKey: "your-api-key",
+});
+
+const client = createSmartAccountClient({
+  feeEstimator: alchemyFeeEstimator(alchemyTransport),
+  ...otherParams,
+});
+```
+
+## Parameters
+
+### transport
+
+`AlchemyTransport`
+An alchemy transport for making Alchemy specific RPC calls
+
+## Returns
+
+`ClientMiddlewareFn`
+A middleware function that takes a transaction structure and fee options, and returns the augmented structure with estimated fees
