+++ docs/pages/reference/account-kit/infra/functions/alchemyGasManagerMiddleware.mdx
@@ -1,0 +1,47 @@
+---
+# This file is autogenerated
+title: alchemyGasManagerMiddleware
+description: Overview of the alchemyGasManagerMiddleware method
+slug: wallets/reference/account-kit/infra/functions/alchemyGasManagerMiddleware
+---
+
+Paymaster middleware factory that uses Alchemy's Gas Manager for sponsoring
+transactions. Adheres to the ERC-7677 standardized communication protocol.
+
+## Import
+
+```ts
+import { alchemyGasManagerMiddleware } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { sepolia, alchemyGasManagerMiddleware } from "@account-kit/infra";
+import { http } from "viem";
+
+const client = createSmartAccountClient({
+  transport: http("rpc-url"),
+  chain: sepolia,
+  ...alchemyGasManagerMiddleware("policyId"),
+});
+```
+
+## Parameters
+
+### policyId
+
+`string | string[]`
+
+- The policyId (or list of policyIds) for Alchemy's gas manager
+
+### policyToken
+
+`PolicyToken | undefined`
+
+- The policy token configuration
+
+## Returns
+
+`Pick<ClientMiddlewareConfig, "dummyPaymasterAndData" | "paymasterAndData">`
+Partial client middleware configuration containing `dummyPaymasterAndData` and `paymasterAndData`
