+++ docs/pages/reference/account-kit/infra/functions/createAlchemySmartAccountClient.mdx
@@ -1,0 +1,38 @@
+---
+# This file is autogenerated
+title: createAlchemySmartAccountClient
+description: Overview of the createAlchemySmartAccountClient method
+slug: wallets/reference/account-kit/infra/functions/createAlchemySmartAccountClient
+---
+
+Creates an Alchemy smart account client using the provided configuration options, including account details, gas manager configuration, and custom middleware.
+
+## Import
+
+```ts
+import { createAlchemySmartAccountClient } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { createAlchemySmartAccountClient, alchemy } from "@account-kit/infra";
+import { sepolia } from "@account-kit/infra/chain";
+
+const client = createAlchemySmartAccountClient({
+  chain: sepolia,
+  transport: alchemy({ apiKey: "your-api-key" }),
+});
+```
+
+## Parameters
+
+### config
+
+`AlchemySmartAccountClientConfig`
+The configuration for creating the Alchemy smart account client
+
+## Returns
+
+`AlchemySmartAccountClient`
+An instance of `AlchemySmartAccountClient` configured based on the provided options
