+++ docs/pages/reference/account-kit/infra/functions/defineAlchemyChain.mdx
@@ -1,0 +1,48 @@
+---
+# This file is autogenerated
+title: defineAlchemyChain
+description: Overview of the defineAlchemyChain method
+slug: wallets/reference/account-kit/infra/functions/defineAlchemyChain
+---
+
+Defines an Alchemy chain configuration by adding an Alchemy-specific RPC base URL to the chain's RPC URLs.
+
+## Import
+
+```ts
+import { defineAlchemyChain } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { defineAlchemyChain } from "@account-kit/infra";
+import { sepolia } from "viem/chains";
+
+const chain = defineAlchemyChain({
+  chain: sepolia,
+  rpcBaseUrl: "https://eth-sepolia.g.alchemy.com/v2",
+});
+```
+
+## Parameters
+
+### params
+
+`AlchemyChainConfig`
+The parameters for defining the Alchemy chain
+
+### params.chain
+
+`Chain`
+The original chain configuration
+
+### params.rpcBaseUrl
+
+`string`
+The Alchemy-specific RPC base URL
+
+## Returns
+
+`Chain`
+The updated chain configuration with the Alchemy RPC URL added
