+++ docs/pages/reference/account-kit/infra/functions/getAlchemyPaymasterAddress.mdx
@@ -1,0 +1,39 @@
+---
+# This file is autogenerated
+title: getAlchemyPaymasterAddress
+description: Overview of the getAlchemyPaymasterAddress method
+slug: wallets/reference/account-kit/infra/functions/getAlchemyPaymasterAddress
+---
+
+Retrieves the Alchemy paymaster address for the given chain. Returns different addresses based on the chain ID.
+
+## Import
+
+```ts
+import { getAlchemyPaymasterAddress } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { sepolia, getAlchemyPaymasterAddress } from "@account-kit/infra";
+
+const paymasterAddress = getAlchemyPaymasterAddress(sepolia, "0.6.0");
+```
+
+## Parameters
+
+### chain
+
+`Chain`
+The chain for which the paymaster address is required
+
+### version
+
+`EntryPointVersion`
+The version of the entry point
+
+## Returns
+
+`Address`
+The Alchemy paymaster address corresponding to the specified chain
