+++ docs/pages/reference/account-kit/infra/functions/getDefaultUserOperationFeeOptions.mdx
@@ -1,0 +1,35 @@
+---
+# This file is autogenerated
+title: getDefaultUserOperationFeeOptions
+description: Overview of the getDefaultUserOperationFeeOptions method
+slug: wallets/reference/account-kit/infra/functions/getDefaultUserOperationFeeOptions
+---
+
+Retrieves the default user operation fee options for a given chain. Adjusts fees for specific chains like Arbitrum and Optimism.
+
+## Import
+
+```ts
+import { getDefaultUserOperationFeeOptions } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { getDefaultUserOperationFeeOptions } from "@account-kit/infra";
+import { arbitrum } from "@account-kit/infra";
+
+const feeOpts = getDefaultUserOperationFeeOptions(arbitrum);
+```
+
+## Parameters
+
+### chain
+
+`Chain`
+The blockchain chain for which to get the fee options
+
+## Returns
+
+`UserOperationFeeOptions`
+An object containing the default fee options for user operations on the specified chain
