+++ docs/pages/reference/account-kit/infra/functions/isAlchemySmartAccountClient.mdx
@@ -1,0 +1,36 @@
+---
+# This file is autogenerated
+title: isAlchemySmartAccountClient
+description: Overview of the isAlchemySmartAccountClient method
+slug: wallets/reference/account-kit/infra/functions/isAlchemySmartAccountClient
+---
+
+Checks if a given client is an Alchemy Smart Account Client. The goal of this check is to ensure that the client supports certain RPC methods.
+
+## Import
+
+```ts
+import { isAlchemySmartAccountClient } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { isAlchemySmartAccountClient } from "@account-kit/infra";
+
+if (isAlchemySmartAccountClient(client)) {
+  // do things with the client as an Alchemy Smart Account Client
+}
+```
+
+## Parameters
+
+### client
+
+`Client<TTransport, TChain, TAccount>`
+The client instance to be checked
+
+## Returns
+
+`boolean`
+`true` if the client is an Alchemy Smart Account Client, otherwise `false`
