+++ docs/pages/reference/account-kit/infra/functions/isAlchemyTransport.mdx
@@ -1,0 +1,33 @@
+---
+# This file is autogenerated
+title: isAlchemyTransport
+description: Overview of the isAlchemyTransport method
+slug: wallets/reference/account-kit/infra/functions/isAlchemyTransport
+---
+
+A type guard for the transport to determine if it is an Alchemy transport.
+Used in cases where we would like to do switching depending on the transport, where there used
+to be two clients for a alchemy and a non alchemy, and with this switch we don't need the two seperate clients. \*
+
+## Import
+
+```ts
+import { isAlchemyTransport } from "@account-kit/infra";
+```
+
+## Parameters
+
+### transport
+
+`Transport`
+The transport to check
+
+### chain
+
+`Chain`
+Chain for the transport to run its function to return the transport config
+
+## Returns
+
+`boolean`
+`true` if the transport is an Alchemy transport, otherwise `false`
