+++ docs/pages/reference/account-kit/infra/functions/simulateUserOperationChanges.mdx
@@ -1,0 +1,42 @@
+---
+# This file is autogenerated
+title: simulateUserOperationChanges
+description: Overview of the simulateUserOperationChanges method
+slug: wallets/reference/account-kit/infra/functions/simulateUserOperationChanges
+---
+
+Simulates user operation changes including asset changes for a specified user operation and returns the resulting state changes.
+
+## Import
+
+```ts
+import { simulateUserOperationChanges } from "@account-kit/infra";
+```
+
+## Usage
+
+```ts
+import { simulateUserOperationChanges, createAlchemyPublicRpcClient } from "@account-kit/infra";
+
+const client = createAlchemyPublicRpcClient(...);
+const response = await simulateUserOperationChanges(client, {
+uo: ...
+});
+```
+
+## Parameters
+
+### client
+
+`Client<Transport, TChain, TAccount, AlchemyRpcSchema>`
+The client instance used to send the simulation request
+
+### args
+
+`SendUserOperationParameters<TAccount>`
+The parameters of the user operation including the account and other overrides
+
+## Returns
+
+`Promise<SimulateUserOperationAssetChangesResponse>`
+A promise that resolves to the response of the simulation showing the asset changes
