+++ packages/signer-web/src/createWebAuthClient.ts
@@ -4,105 +4,6 @@ import { IframeStamper } from "@turnkey/iframe-stamper";
 const CHECK_CLOSE_INTERVAL = 500;
 
 /**
- * Reads and removes the specified query params from the URL.
- *
- * @param {T} keys object whose values are the query parameter keys to read and
- * remove
- * @returns {{ [K in keyof T]: string | undefined }} object with the same keys
- * as the input whose values are the values of the query params.
- */
-function getAndRemoveQueryParams<T extends Record<string, string>>(
-  keys: T
-): { [K in keyof T]: string | undefined } {
-  const url = new URL(window.location.href);
-  const result: Record<string, string | undefined> = {};
-  let foundQueryParam = false;
-  for (const [key, param] of Object.entries(keys)) {
-    const value = url.searchParams.get(param) ?? undefined;
-    foundQueryParam ||= value != null;
-    result[key] = value;
-    url.searchParams.delete(param);
-  }
-  if (foundQueryParam) {
-    window.history.replaceState(window.history.state, "", url.toString());
-  }
-  return result as { [K in keyof T]: string | undefined };
-}
-
-/**
- * Attempts to extract OAuth callback parameters from the current URL.
- * Returns the extracted parameters if this appears to be an OAuth callback,
- * or null if no OAuth parameters are found.
- *
- * @returns {object | null} OAuth callback parameters or null if not a callback
- */
-function extractOAuthCallbackParams() {
-  const qpStructure = {
-    status: "alchemy-status",
-    oauthBundle: "alchemy-bundle",
-    oauthOrgId: "alchemy-org-id",
-    idToken: "alchemy-id-token",
-    isSignup: "aa-is-signup",
-    otpId: "alchemy-otp-id",
-    email: "alchemy-email",
-    authProvider: "alchemy-auth-provider",
-    oauthError: "alchemy-error",
-  };
-
-  const {
-    status,
-    oauthBundle,
-    oauthOrgId,
-    idToken,
-    isSignup,
-    otpId,
-    email,
-    authProvider,
-    oauthError,
-  } = getAndRemoveQueryParams(qpStructure);
-
-  // Check if this is an OAuth callback by looking for required OAuth parameters
-  if (oauthBundle && oauthOrgId && idToken) {
-    return {
-      status: "SUCCESS" as const,
-      bundle: oauthBundle,
-      orgId: oauthOrgId,
-      idToken,
-      isSignup: isSignup === "true",
-    };
-  }
-
-  // Check for OAuth error
-  if (oauthError) {
-    return {
-      status: "ERROR" as const,
-      error: oauthError,
-    };
-  }
-
-  // Check for account linking prompt
-  if (
-    status === "ACCOUNT_LINKING_CONFIRMATION_REQUIRED" &&
-    idToken &&
-    email &&
-    authProvider &&
-    otpId &&
-    oauthOrgId
-  ) {
-    return {
-      status: "ACCOUNT_LINKING_CONFIRMATION_REQUIRED" as const,
-      idToken,
-      email,
-      providerName: authProvider,
-      otpId,
-      orgId: oauthOrgId,
-    };
-  }
-
-  return null;
-}
-
-/**
  * Configuration parameters for creating a web-based AuthClient
  */
 export type WebAuthClientParams = {
@@ -266,16 +167,6 @@ export function createWebAuthClient({
           });
 
         case "redirect":
-          // // First, check if we're currently on a page with OAuth callback parameters
-          // const callbackParams = extractOAuthCallbackParams();
-          // if (callbackParams) {
-          //   // We're on the OAuth callback - return the extracted parameters
-          //   if (callbackParams.status === "ERROR") {
-          //     throw new Error(callbackParams.error);
-          //   }
-          //   return callbackParams;
-          // }
-
           // No OAuth callback detected, so initiate the redirect
           window.location.href = authUrl;
           return new Promise((_, reject) =>
