+++ packages/signer/src/authClient.ts
@@ -70,6 +70,105 @@ type PendingOtp = {
 };
 
 /**
+ * Reads and removes the specified query params from the URL.
+ *
+ * @param {T} keys object whose values are the query parameter keys to read and
+ * remove
+ * @returns {{ [K in keyof T]: string | undefined }} object with the same keys
+ * as the input whose values are the values of the query params.
+ */
+function getAndRemoveQueryParams<T extends Record<string, string>>(
+  keys: T
+): { [K in keyof T]: string | undefined } {
+  const url = new URL(window.location.href);
+  const result: Record<string, string | undefined> = {};
+  let foundQueryParam = false;
+  for (const [key, param] of Object.entries(keys)) {
+    const value = url.searchParams.get(param) ?? undefined;
+    foundQueryParam ||= value != null;
+    result[key] = value;
+    url.searchParams.delete(param);
+  }
+  if (foundQueryParam) {
+    window.history.replaceState(window.history.state, "", url.toString());
+  }
+  return result as { [K in keyof T]: string | undefined };
+}
+
+/**
+ * Attempts to extract OAuth callback parameters from the current URL.
+ * Returns the extracted parameters if this appears to be an OAuth callback,
+ * or null if no OAuth parameters are found.
+ *
+ * @returns {object | null} OAuth callback parameters or null if not a callback
+ */
+function extractOAuthCallbackParams() {
+  const qpStructure = {
+    status: "alchemy-status",
+    oauthBundle: "alchemy-bundle",
+    oauthOrgId: "alchemy-org-id",
+    idToken: "alchemy-id-token",
+    isSignup: "aa-is-signup",
+    otpId: "alchemy-otp-id",
+    email: "alchemy-email",
+    authProvider: "alchemy-auth-provider",
+    oauthError: "alchemy-error",
+  };
+
+  const {
+    status,
+    oauthBundle,
+    oauthOrgId,
+    idToken,
+    isSignup,
+    otpId,
+    email,
+    authProvider,
+    oauthError,
+  } = getAndRemoveQueryParams(qpStructure);
+
+  // Check if this is an OAuth callback by looking for required OAuth parameters
+  if (oauthBundle && oauthOrgId && idToken) {
+    return {
+      status: "SUCCESS" as const,
+      bundle: oauthBundle,
+      orgId: oauthOrgId,
+      idToken,
+      isSignup: isSignup === "true",
+    };
+  }
+
+  // Check for OAuth error
+  if (oauthError) {
+    return {
+      status: "ERROR" as const,
+      error: oauthError,
+    };
+  }
+
+  // Check for account linking prompt
+  if (
+    status === "ACCOUNT_LINKING_CONFIRMATION_REQUIRED" &&
+    idToken &&
+    email &&
+    authProvider &&
+    otpId &&
+    oauthOrgId
+  ) {
+    return {
+      status: "ACCOUNT_LINKING_CONFIRMATION_REQUIRED" as const,
+      idToken,
+      email,
+      providerName: authProvider,
+      otpId,
+      orgId: oauthOrgId,
+    };
+  }
+
+  return null;
+}
+
+/**
  * AuthClient handles authentication flows including email OTP, OAuth, and passkey authentication.
  * This is a simplified authentication client that provides methods for different authentication types.
  *
@@ -230,6 +329,50 @@ export class AuthClient {
     } else {
       throw new Error(`Unknown OAuth flow response: ${response.status}`);
     }
+  }
+
+  /**
+   * Handles OAuth redirect callback by extracting OAuth parameters from the current URL.
+   * This method should be called when the user returns to your application after completing
+   * OAuth authentication with a provider (e.g., Google, Facebook).
+   *
+   * The method automatically detects OAuth callback parameters (alchemy-bundle, alchemy-org-id,
+   * alchemy-id-token) in the URL query string, processes them, and completes the authentication flow.
+   *
+   * @returns {Promise<Signer>} Promise that resolves to an authenticated Signer instance
+   * @throws {Error} If no OAuth callback parameters are found in the URL
+   * @throws {Error} If OAuth authentication failed (alchemy-error parameter present)
+   * @throws {Error} If credential bundle injection fails
+   *
+   * @example
+   * ```ts
+   * // After user returns from OAuth provider with callback URL like:
+   * // https://yourapp.com/?alchemy-bundle=eyJ0eXAi...&alchemy-org-id=24c1ac...&alchemy-id-token=eyJ0eXAi...
+   *
+   * try {
+   *   const signer = await authClient.handleOauthRedirect();
+   *   console.log("OAuth authentication successful", signer.user);
+   * } catch (error) {
+   *   console.error("OAuth callback failed:", error.message);
+   * }
+   * ```
+   */
+  public async handleOauthRedirect(): Promise<Signer | null> {
+    // // First, check if we're currently on a page with OAuth callback parameters
+    const callbackParams = extractOAuthCallbackParams();
+    if (callbackParams) {
+      // We're on the OAuth callback - return the extracted parameters
+      if (callbackParams.status === "ERROR") {
+        throw new Error(callbackParams.error);
+      }
+      return this.completeAuthWithBundle({
+        bundle: callbackParams.bundle!,
+        orgId: callbackParams.orgId!,
+        idToken: callbackParams.idToken,
+      });
+    }
+    console.log("No OAuth callback parameters found");
+    return null;
   }
 
   /**
